+++ airflow/providers/qubole/operators/qubole.py
@@ -33,7 +33,7 @@ from airflow.providers.qubole.hooks.qubole import (
 
 if TYPE_CHECKING:
 
-    from airflow.models.taskinstance import TaskInstanceKey
+    from airflow.models.taskinstancekey import TaskInstanceKey
     from airflow.utils.context import Context
